Sprint 11: canonicalize Phase 2 gate runners
