+++ .ai/handoff/CURRENT_STATE.md
@@ -3,7 +3,8 @@
 ## Canonical Truth
 
 - The working repo state is current through Phase 2 Sprint 7.
-- The accepted baseline includes deterministic Phase 2 gate tooling: `python3 scripts/run_phase2_acceptance.py`, `python3 scripts/run_phase2_readiness_gates.py`, and `python3 scripts/run_phase2_validation_matrix.py` (default go/no-go command). MVP script names remain supported aliases with identical semantics.
+- The accepted baseline includes deterministic Phase 2 gate tooling: `python3 scripts/run_phase2_acceptance.py`, `python3 scripts/run_phase2_readiness_gates.py`, and `python3 scripts/run_phase2_validation_matrix.py` (default go/no-go command).
+- Gate ownership is canonicalized to Phase 2 runner scripts; `run_mvp_acceptance.py`, `run_mvp_readiness_gates.py`, and `run_mvp_validation_matrix.py` remain supported compatibility aliases with identical semantics.
 - Use [PRODUCT_BRIEF.md](../../PRODUCT_BRIEF.md) for product scope, [ARCHITECTURE.md](../../ARCHITECTURE.md) for implemented boundaries, [ROADMAP.md](../../ROADMAP.md) for forward planning, and [RULES.md](../../RULES.md) for durable operating rules.
 - The live sprint reports are [BUILD_REPORT.md](../../BUILD_REPORT.md) and [REVIEW_REPORT.md](../../REVIEW_REPORT.md) at repo root; older accepted sprint history belongs in [docs/archive/sprints](../../docs/archive/sprints), not in this handoff.
