+++ BUILD_REPORT.md
@@ -1,61 +1,85 @@
 # BUILD_REPORT.md
 
 ## Sprint Objective
-Add deterministic acceptance evidence that explicit-signal capture writes flow into resumption-brief context and remain included in the Phase 2 acceptance gate chain.
+Canonicalize Phase 2 gate runners as the single implementation source, convert MVP runners to explicit compatibility aliases, and preserve deterministic gate semantics/thresholds unchanged.
 
 ## Completed Work
-- Added new deterministic acceptance scenario:
-  - `tests/integration/test_mvp_acceptance_suite.py::test_acceptance_explicit_signal_capture_flows_into_resumption_brief`
-- Scenario setup seeds one thread with two `message.user` explicit-signal events (preference + commitment) and one assistant event.
-- Scenario validates `POST /v0/memories/capture-explicit-signals` outcomes for both signals:
-  - preference capture: candidate payload shape, `ADD` admission, and aggregate summary counts.
-  - commitment capture: candidate payload shape, `ADD` admission, `open_loop.decision == "CREATED"`, and aggregate summary counts.
-- Scenario validates continuity into `GET /v0/threads/{thread_id}/resumption-brief`:
-  - open-loop section summary (`limit/returned_count/total_count/order`) and exact item continuity with created open-loop id/memory id/title.
-  - memory-highlights section summary (`limit/returned_count/total_count/order`) and deterministic keys/values/source_event_ids for both captured memories.
-- Updated `scripts/run_mvp_acceptance.py` to include the new scenario in gate execution.
-- Confirmed `scripts/run_phase2_acceptance.py` executes the updated acceptance chain unchanged (no code changes required in alias script).
-
-## Exact Runner Command List Delta
-- Added scenario key in `ACCEPTANCE_SCENARIOS`:
-  - `capture_resumption`
-- Added acceptance node id in `ACCEPTANCE_TEST_NODE_IDS`:
-  - `tests/integration/test_mvp_acceptance_suite.py::test_acceptance_explicit_signal_capture_flows_into_resumption_brief`
-
-## Explicit Assertions Proving Capture-To-Resumption Continuity
-- Capture endpoint assertions prove persistence/admission outcomes:
-  - preference candidate key/value/source metadata + summary counts.
-  - commitment candidate key/value/open-loop title + `ADD` admission + `open_loop CREATED` summary counts.
-- Resumption-brief assertions prove downstream continuity:
-  - open-loop item equals created loop identity and title from capture output.
-  - memory-highlights include both captured memory keys in deterministic order with expected values and source event ids.
+- Migrated canonical gate ownership to Phase 2 scripts:
+  - `scripts/run_phase2_acceptance.py`
+  - `scripts/run_phase2_readiness_gates.py`
+  - `scripts/run_phase2_validation_matrix.py`
+- Removed reverse delegation from Phase 2 scripts to MVP scripts.
+- Converted MVP scripts to explicit compatibility aliases that forward args/exit codes to Phase 2 scripts:
+  - `scripts/run_mvp_acceptance.py -> scripts/run_phase2_acceptance.py`
+  - `scripts/run_mvp_readiness_gates.py -> scripts/run_phase2_readiness_gates.py`
+  - `scripts/run_mvp_validation_matrix.py -> scripts/run_phase2_validation_matrix.py`
+- Preserved deterministic parity (no semantic widening):
+  - acceptance scenario node list unchanged
+  - gate thresholds unchanged (`latency_p95 < 5.0`, `cache_reuse >= 0.70`, `memory precision > 0.80` with `adjudicated_sample >= 20`)
+  - no-go behavior unchanged (`FAIL`/`BLOCKED` returns non-zero)
+- Updated unit verification for canonical direction and deterministic wiring in `tests/unit/test_phase2_gate_wrappers.py`:
+  - MVP alias target mapping
+  - arg/exit-code forwarding behavior
+  - python resolver behavior
+  - assertions that Phase 2 scripts do not call `run_mvp_*`
+  - assertions that readiness/validation wiring uses Phase 2 command chain
+- Updated sprint-scoped docs to state canonical ownership clearly:
+  - `docs/runbooks/mvp-acceptance-suite.md`
+  - `docs/runbooks/mvp-readiness-gates.md`
+  - `docs/runbooks/mvp-validation-matrix.md`
+  - `README.md`
+  - `.ai/handoff/CURRENT_STATE.md`
 
 ## Incomplete Work
 - None within sprint scope.
 
 ## Files Changed
-- `tests/integration/test_mvp_acceptance_suite.py`
+- `scripts/run_phase2_acceptance.py`
+- `scripts/run_phase2_readiness_gates.py`
+- `scripts/run_phase2_validation_matrix.py`
 - `scripts/run_mvp_acceptance.py`
+- `scripts/run_mvp_readiness_gates.py`
+- `scripts/run_mvp_validation_matrix.py`
+- `tests/unit/test_phase2_gate_wrappers.py`
+- `docs/runbooks/mvp-acceptance-suite.md`
+- `docs/runbooks/mvp-readiness-gates.md`
+- `docs/runbooks/mvp-validation-matrix.md`
+- `README.md`
+- `.ai/handoff/CURRENT_STATE.md`
 - `BUILD_REPORT.md`
 - `REVIEW_REPORT.md`
 
 ## Tests Run
-1. `PYTHONPATH=$PWD .venv/bin/pytest tests/integration/test_mvp_acceptance_suite.py::test_acceptance_explicit_signal_capture_flows_into_resumption_brief`
-- Outcome: PASS (`1 passed`, exit code `0`).
+1. `./.venv/bin/python -m pytest tests/unit/test_phase2_gate_wrappers.py`
+- Outcome: PASS (`17 passed`, exit code `0`).
 
 2. `python3 scripts/run_phase2_acceptance.py`
-- Outcome: PASS (`4 passed`, exit code `0`).
-- Includes updated acceptance chain containing the new scenario node id.
+- Initial sandboxed run: blocked by local Postgres access (`psycopg.OperationalError: Operation not permitted`).
+- Escalated verification run: PASS (`4 passed`, exit code `0`).
+
+3. `python3 scripts/run_phase2_readiness_gates.py --induce-gate acceptance_fail`
+- Outcome: expected NO_GO (exit code `1`).
+- Gate outcomes:
+  - `acceptance_suite`: `FAIL` (induced)
+  - `latency_p95`: `PASS`
+  - `cache_reuse`: `PASS`
+  - `memory_quality`: `PASS`
+
+4. `python3 scripts/run_phase2_validation_matrix.py --induce-step readiness_gates`
+- Outcome: expected NO_GO (exit code `1`).
+- Step outcomes:
+  - `readiness_gates`: `FAIL` (induced, exit code `97`)
+  - `backend_integration_matrix`: `PASS` (`76 passed`)
+  - `web_validation_matrix`: `PASS` (`13 files, 64 tests passed`)
 
 ## Blockers/Issues
-- Integration tests require local Postgres connectivity; sandboxed execution initially failed with connection permission errors.
-- Verification completed successfully after running the required commands with escalated permissions.
-- No API/contract changes were required.
+- No implementation blockers.
+- Verification requiring local Postgres cannot run in default sandbox; escalated execution was required for full sprint command evidence.
 
 ## Deferred Scope (Explicit)
-- Automation implementation: deferred (out of scope).
-- Workers/orchestration implementation: deferred (out of scope).
-- Phase 3 runtime/profile routing or orchestration: deferred (out of scope).
+- Automation/workers/orchestration implementation: deferred (out of scope).
+- Phase 3 runtime/profile routing: deferred (out of scope).
+- Any API/endpoint/schema behavior expansion: deferred (out of scope).
 
 ## Recommended Next Step
-Proceed to Control Tower integration review to confirm sprint-scope adherence and acceptance-hardening evidence, then advance the sprint PR.
+Proceed to Control Tower integration review to confirm sprint-scope adherence and approve Phase 2 canonical gate ownership with MVP compatibility aliases.
