+++ REVIEW_REPORT.md
@@ -4,26 +4,35 @@
 PASS
 
 ## criteria met
-- Sprint stayed capture-to-resumption acceptance scoped (`tests/integration/test_mvp_acceptance_suite.py`, `scripts/run_mvp_acceptance.py`, sprint reports).
-- Deterministic acceptance scenario was added: `tests/integration/test_mvp_acceptance_suite.py::test_acceptance_explicit_signal_capture_flows_into_resumption_brief`.
-- Scenario proves required chain end-to-end: `message.user` explicit preference/commitment events are seeded; `POST /v0/memories/capture-explicit-signals` assertions cover candidates, admissions, and summary counters; `GET /v0/threads/{thread_id}/resumption-brief` assertions cover open-loop continuity and memory-highlight continuity.
-- Acceptance runner includes the new scenario: `scripts/run_mvp_acceptance.py` adds scenario key `capture_resumption` and node id `...::test_acceptance_explicit_signal_capture_flows_into_resumption_brief`.
-- Phase 2 alias behavior is preserved: `scripts/run_phase2_acceptance.py` remains an alias to `scripts/run_mvp_acceptance.py` and executes updated chain.
-- Verification commands succeeded in reviewer run: `PYTHONPATH=$PWD .venv/bin/pytest tests/integration/test_mvp_acceptance_suite.py::test_acceptance_explicit_signal_capture_flows_into_resumption_brief` passed (`1 passed`), and `python3 scripts/run_phase2_acceptance.py` passed (`4 passed`).
-- No endpoint/product contract changes were introduced to satisfy this sprint.
+- `scripts/run_phase2_acceptance.py`, `scripts/run_phase2_readiness_gates.py`, and `scripts/run_phase2_validation_matrix.py` are now canonical implementations and no longer delegate to `run_mvp_*` scripts.
+- `scripts/run_mvp_acceptance.py`, `scripts/run_mvp_readiness_gates.py`, and `scripts/run_mvp_validation_matrix.py` are explicit compatibility aliases that forward args and exit codes to the Phase 2 scripts.
+- Deterministic gate semantics were preserved during migration:
+  - acceptance node list is unchanged (moved from MVP runner to Phase 2 runner)
+  - readiness thresholds are unchanged (`latency_p95 < 5.0`, `cache_reuse >= 0.70`, `memory precision > 0.80`, `adjudicated_sample >= 20`)
+  - validation-matrix step ordering and no-go behavior remain unchanged
+- `tests/unit/test_phase2_gate_wrappers.py` was updated for canonical direction and passed in reviewer execution:
+  - command run: `./.venv/bin/python -m pytest tests/unit/test_phase2_gate_wrappers.py -q`
+  - result: `17 passed`
+- Runbooks and entry docs were aligned to “Phase 2 canonical, MVP compatibility alias” across:
+  - `docs/runbooks/mvp-acceptance-suite.md`
+  - `docs/runbooks/mvp-readiness-gates.md`
+  - `docs/runbooks/mvp-validation-matrix.md`
+  - `README.md`
+  - `.ai/handoff/CURRENT_STATE.md`
+- No product endpoint/API surface changes were introduced in this sprint diff.
 
 ## criteria missed
 - None.
 
 ## quality issues
-- None blocking found.
+- None blocking.
 
 ## regression risks
 - Low.
-- Residual risk: future drift between acceptance scenario list and suite node ids; current wiring is consistent and passing.
+- Residual risk: if future changes update only canonical script messaging or only alias messaging, operator-facing output may drift even if execution semantics stay stable.
 
 ## docs issues
-- None blocking.
+- None blocking within sprint scope.
 
 ## should anything be added to RULES.md?
 - No.
@@ -32,5 +41,5 @@ PASS
 - No.
 
 ## recommended next action
-1. Proceed to Control Tower approval with this sprint marked `PASS`.
-2. Keep `run_mvp_acceptance.py` scenario/node-id lists synchronized as new acceptance scenarios are added in future sprints.
+1. Approve sprint as complete and proceed with Control Tower integration sign-off.
+2. In future gate changes, update Phase 2 scripts first and keep MVP alias scripts thin and argument-transparent.
