+++ docs/runbooks/mvp-acceptance-suite.md
@@ -2,6 +2,7 @@
 
 ## Objective
 Run one deterministic acceptance suite that validates shipped MVP-critical journeys and returns a single reviewer-ready pass/fail signal.
+Canonical implementation is Phase 2 (`run_phase2_acceptance.py`); `run_mvp_acceptance.py` is a compatibility alias.
 
 ## Prerequisites
 - Local dependencies installed (`python3 -m venv .venv` and `./.venv/bin/python -m pip install -e '.[dev]'`).
@@ -12,6 +13,8 @@ Run one deterministic acceptance suite that validates shipped MVP-critical journ
 - `response_memory`:
   - context-aware response uses admitted memory evidence
   - preference correction is reflected in subsequent compile + response flow
+- `capture_resumption`:
+  - explicit-signal capture writes propagate into thread resumption-brief continuity
 - `approval_execution`:
   - approval-required request lifecycle (pending -> approved -> executed) stays linked
   - consequential trace evidence remains queryable
@@ -20,7 +23,7 @@ Run one deterministic acceptance suite that validates shipped MVP-critical journ
 
 ## Exact Command (Normal)
 ```bash
-python3 scripts/run_mvp_acceptance.py
+python3 scripts/run_phase2_acceptance.py
 ```
 
 Expected behavior:
@@ -33,18 +36,28 @@ Expected behavior:
 
 ## Exact Command (Induced-Failure Check)
 ```bash
-python3 scripts/run_mvp_acceptance.py --induce-failure approval_execution
+python3 scripts/run_phase2_acceptance.py --induce-failure approval_execution
 ```
 
 Expected behavior:
 - Intentionally fails exactly the selected scenario via `MVP_ACCEPTANCE_INDUCED_FAILURE_SCENARIO`.
-- Returns non-zero exit code and prints `MVP acceptance suite result: FAIL (...)`.
+- Returns non-zero exit code and prints `Phase 2 acceptance suite result: FAIL (...)`.
 - This validates deterministic negative signaling for reviewers.
 
 Other valid induced-failure scenario names:
 - `response_memory`
+- `capture_resumption`
 - `approval_execution`
 - `magnesium_reorder`
+
+## Compatibility Alias Command
+```bash
+python3 scripts/run_mvp_acceptance.py
+```
+
+Expected behavior:
+- Prints explicit alias messaging and delegates to `scripts/run_phase2_acceptance.py`.
+- Preserves the same arguments, thresholds, scenario coverage, and exit semantics.
 
 ## Optional Direct Pytest Command
 ```bash
